Rename Teams AI to Teams SDK 🎉 (#2671)
Renames 'Teams AI' to 'Teams SDK' comprehensively throughout this repo.
This includes documentation, comments, and code snippets.
- Change references to Teams AI to Teams SDK
- Limit use of "Teams Library" in favor of "Teams SDK". Some uses of the
word library were kept in cases of clarity, or when not referencing this
SDK (e.g. MSAL)
- Limit use of 'v2' throughout the repo. Exceptions being root README.md
for v1/v2 findability, migrations documentation, and in the few
instances where differentiation of the two was still needed.
- Announcement bar set in docs, announcement set in root README.md
- references to the GitHub repo itself, including the monorepo's package
name
- github.io / base-url reference
- URL links that already use teams-ai
- package name references

---------

Co-authored-by: Corina Gum <>